Implement showing of git branch in the prompt
We keep track of if we are inside a git repository inside the ShellState
struct as well as if the branch was updated in the last command. In the main
loop we then update the branch name if it wasn't already updated, and then show
the branch in the prompt if we are inside a repository.

We only visit `.git/HEAD` once per `set_cwd()` call and we only run again in
the main loop if we did not run in that iteration yet.

However I discovered that the `set_cwd()` call is being called twice in every
iteration for some reason, and so we also read `.git/HEAD` twice. This is a
separate issue so it should be fixed in a separate PR once this is merged.